Use the event datetime
It is aware, instead of a time tuple to compare properly the modification file
with the event start time and write the file name usign the local time.
This also fixes a problem in which using an icalendar, the timetuple for
the first file was in utc and the rest in local time.
Closes #1465
reviewed by: Marc-André Lureau
Backport ac54e6.